+++ src/main/java/svenhjol/charm/module/block_of_gunpowder/BlockOfGunpowder.java
@@ -10,6 +10,7 @@ import svenhjol.charm.annotation.CommonModule;
 import svenhjol.charm.helper.PlayerHelper;
 import svenhjol.charm.init.CharmAdvancements;
 import svenhjol.charm.loader.CharmModule;
+import svenhjol.charm.module.extra_wandering_trades.ExtraWanderingTrades;
 import svenhjol.charm.registry.CommonRegistry;
 
 @CommonModule(mod = Charm.MOD_ID, description = "A storage block for gunpowder. It obeys gravity and dissolves in lava.")
@@ -22,6 +23,11 @@ public class BlockOfGunpowder extends CharmModule {
     public void register() {
         GUNPOWDER_BLOCK = new GunpowderBlock(this);
         GUNPOWDER_DISSOLVE_SOUND = CommonRegistry.sound(new ResourceLocation(Charm.MOD_ID, "gunpowder_dissolve"));
+    }
+
+    @Override
+    public void runWhenEnabled() {
+        ExtraWanderingTrades.registerItem(GUNPOWDER_BLOCK, 1, 5);
     }
 
     public static void triggerAdvancementForNearbyPlayers(ServerLevel level, BlockPos pos) {