+++ src/main/java/svenhjol/charm/module/player_pressure_plates/PlayerPressurePlates.java
@@ -3,6 +3,7 @@ package svenhjol.charm.module.player_pressure_plates;
 import svenhjol.charm.Charm;
 import svenhjol.charm.annotation.CommonModule;
 import svenhjol.charm.loader.CharmModule;
+import svenhjol.charm.module.extra_wandering_trades.ExtraWanderingTrades;
 
 @CommonModule(mod = Charm.MOD_ID, description = "Player-only pressure plates crafted using gilded blackstone.")
 public class PlayerPressurePlates extends CharmModule {
@@ -11,5 +12,10 @@ public class PlayerPressurePlates extends CharmModule {
     @Override
     public void register() {
         PLAYER_PRESSURE_PLATE_BLOCK = new PlayerPressurePlateBlock(this);
+    }
+
+    @Override
+    public void runWhenEnabled() {
+        ExtraWanderingTrades.registerItem(PLAYER_PRESSURE_PLATE_BLOCK, 2, 8);
     }
 }