+++ src/main/java/svenhjol/charm/module/redstone_lanterns/RedstoneLanterns.java
@@ -3,6 +3,7 @@ package svenhjol.charm.module.redstone_lanterns;
 import svenhjol.charm.Charm;
 import svenhjol.charm.annotation.CommonModule;
 import svenhjol.charm.loader.CharmModule;
+import svenhjol.charm.module.extra_wandering_trades.ExtraWanderingTrades;
 
 @CommonModule(mod = Charm.MOD_ID, description = "A lantern that emits light when a redstone signal is received.")
 public class RedstoneLanterns extends CharmModule {
@@ -11,5 +12,10 @@ public class RedstoneLanterns extends CharmModule {
     @Override
     public void register() {
         REDSTONE_LANTERN = new RedstoneLanternBlock(this);
+    }
+
+    @Override
+    public void runWhenEnabled() {
+        ExtraWanderingTrades.registerItem(REDSTONE_LANTERN, 1, 5);
     }
 }