+++ src/main/java/svenhjol/charm/module/smooth_glowstone/SmoothGlowstone.java
@@ -1,9 +1,9 @@
 package svenhjol.charm.module.smooth_glowstone;
 
-import net.minecraft.resources.ResourceLocation;
 import svenhjol.charm.Charm;
 import svenhjol.charm.annotation.CommonModule;
 import svenhjol.charm.loader.CharmModule;
+import svenhjol.charm.module.extra_wandering_trades.ExtraWanderingTrades;
 
 @CommonModule(mod = Charm.MOD_ID, description = "Smooth Glowstone")
 public class SmoothGlowstone extends CharmModule {
@@ -12,5 +12,10 @@ public class SmoothGlowstone extends CharmModule {
     @Override
     public void register() {
         SMOOTH_GLOWSTONE = new SmoothGlowstoneBlock(this);
+    }
+
+    @Override
+    public void runWhenEnabled() {
+        ExtraWanderingTrades.registerItem(SMOOTH_GLOWSTONE, 2, 5);
     }
 }